modify job controller by remove functions from endpoint and create helper file and add these function in
